refactor(agents): rename wiki_updater to nl_updater, extract shared common module
- common.py: IRRELEVANT_SENTINEL, NO_CHANGE_SENTINEL, strip_outer_fence
- nl_updater.py: process_instruction (NL/MCP path only)
- Remove wiki_updater.py and local _strip_outer_fence copy in batch reconciler
