+++ backend/app/llm/agents/nl_updater.py
@@ -1,4 +1,4 @@
-"""Wiki updater agent — decides whether and how to update a wiki page.
+"""NL instruction updater agent — applies a natural-language instruction to a wiki page.
 
 Single-shot LLM call. The system prompt constrains the output to either
 the literal token ``NO_CHANGE`` or the full new page body in markdown —
@@ -10,13 +10,11 @@ import logging
 from typing import Any
 
 from app.llm import client
+from app.llm.agents.common import NO_CHANGE_SENTINEL, strip_outer_fence
 from app.llm.prompts import load_prompt
 from app.tracing import trace_flow
 
 log = logging.getLogger(__name__)
-
-NO_CHANGE_SENTINEL = "NO_CHANGE"
-IRRELEVANT_SENTINEL = "IRRELEVANT"
 
 
 def process_instruction(wiki_path: str, current_body: str, payload: dict[str, Any], source: str) -> str | None:
@@ -35,7 +33,7 @@ def process_instruction(wiki_path: str, current_body: str, payload: dict[str, An
         current_body=current_body,
         payload=payload,
     )
-    with trace_flow("agent.wiki_updater", wiki_path=wiki_path, source=source):
+    with trace_flow("agent.nl_updater", wiki_path=wiki_path, source=source):
         result = client.complete(
             messages=[
                 {"role": "system", "content": system},
@@ -44,26 +42,8 @@ def process_instruction(wiki_path: str, current_body: str, payload: dict[str, An
         )
     text = result.text.strip()
     if not text:
-        log.warning("wiki_updater returned empty text for %s", wiki_path)
+        log.warning("nl_updater returned empty text for %s", wiki_path)
         return None
     if text == NO_CHANGE_SENTINEL or text.startswith(NO_CHANGE_SENTINEL + "\n"):
         return None
-    # Defensive: strip a single leading/trailing markdown fence if the
-    # model added one despite the prompt. Don't strip nested fences —
-    # those are part of the body.
-    return _strip_outer_fence(text)
-
-
-
-def _strip_outer_fence(text: str) -> str:
-    if not text.startswith("```"):
-        return text
-    first_nl = text.find("\n")
-    if first_nl == -1:
-        return text
-    if not text.rstrip().endswith("```"):
-        return text
-    inner = text[first_nl + 1 :].rstrip()
-    if inner.endswith("```"):
-        inner = inner[: -3].rstrip()
-    return inner + "\n" if text.endswith("\n") else inner
+    return strip_outer_fence(text)
