+++ backend/tests/test_ingest_pipeline.py
@@ -21,7 +21,7 @@ from app.config import CONFIG
 from app.db.fts import SearchHit
 from app.ingest import search as ingest_search
 from app.ingest.source_tiers import is_filtered
-from app.llm.agents.wiki_updater import IRRELEVANT_SENTINEL
+from app.llm.agents.common import IRRELEVANT_SENTINEL
 from app.llm.settings import _EMPTY as _EMPTY_LLM_SETTINGS, LLMSettings
